feat: introduce cart system

+++ src/domain/product/view.rs
@@ -1,4 +1,5 @@
 use chrono::{DateTime, Utc};
+use rust_decimal::Decimal;
 use serde::{Deserialize, Serialize};
 
 // The view for a ProductView query, for a standard http application this should
@@ -8,7 +9,7 @@ pub struct ProductView {
     #[serde(skip_serializing_if = "Option::is_none")]
     pub id: Option<String>,
     pub name: String,
-    pub price: f32,
+    pub price: Decimal,
     pub description: String,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub created_at: Option<DateTime<Utc>>,
